#193046 - add waitForWorkbenchRestored method to driver - move driver to workbench - wait for workbench to restore if command is not found
