+++ src/vs/workbench/services/driver/electron-sandbox/driver.ts
@@ -3,10 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { BrowserWindowDriver } from 'vs/platform/driver/browser/driver';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 import { IFileService } from 'vs/platform/files/common/files';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { BrowserWindowDriver } from 'vs/workbench/services/driver/browser/driver';
+import { ILifecycleService } from 'vs/workbench/services/lifecycle/common/lifecycle';
 
 interface INativeWindowDriverHelper {
 	exitApplication(): Promise<void>;
@@ -17,9 +18,10 @@ class NativeWindowDriver extends BrowserWindowDriver {
 	constructor(
 		private readonly helper: INativeWindowDriverHelper,
 		@IFileService fileService: IFileService,
-		@IEnvironmentService environmentService: IEnvironmentService
+		@IEnvironmentService environmentService: IEnvironmentService,
+		@ILifecycleService lifecycleService: ILifecycleService
 	) {
-		super(fileService, environmentService);
+		super(fileService, environmentService, lifecycleService);
 	}
 
 	override exitApplication(): Promise<void> {
