+++ test/automation/tools/copy-driver-definition.js
@@ -10,7 +10,7 @@ const fs = require('fs');
 const path = require('path');
 
 const root = path.dirname(path.dirname(path.dirname(__dirname)));
-const driverPath = path.join(root, 'src/vs/platform/driver/common/driver.ts');
+const driverPath = path.join(root, 'src/vs/workbench/services/driver/common/driver.ts');
 
 let contents = fs.readFileSync(driverPath, 'utf8');
 contents = /\/\/\*START([\s\S]*)\/\/\*END/mi.exec(contents)[1].trim();
